Refactor redis tests
