+++ misk-redis/src/test/kotlin/misk/redis/RealRedisTest.kt
@@ -10,7 +10,6 @@ import misk.testing.MiskTest
 import misk.testing.MiskTestModule
 import okio.ByteString.Companion.encodeUtf8
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import redis.clients.jedis.ConnectionPoolConfig
 import wisp.deployment.TESTING
@@ -29,10 +28,10 @@ class RealRedisTest : AbstractRedisTest() {
 
   @Inject override lateinit var redis: Redis
 
-  @BeforeEach
-  fun setUp() {
-    redis.flushAll()
-  }
+  // The following tests are not part of the AbstractRedisTest because as implemented right now,
+  // Redis transactions are not supported in Cluster mode. This is because the Redis interface
+  // leaks Jedis implementation details, and exposes the wrong type for Transactions when
+  // Jedis is configured to use a cluster.
 
   @Test
   fun `watch and unwatch succeeds`() {
